add episode detail api
